Move server picker into chat title
