+++ web/src/shell/AppShell.tsx
@@ -1599,11 +1599,8 @@ export function AppShell() {
                 />
               </div>
             )}
-            {/* The server picker is NOT here: it lives at the bottom of the
-          sidebar (SidebarServerPicker). This strip is shared with the chat
-          header — which is taller and also anchored at top-0 — so a centered
-          "<thread> — <host>" label here collided with the header's action
-          cluster on a narrow window. The strip stays pure drag surface. */}
+            {/* The server picker lives in ChatHeader's chat-relative center
+          column; this window-wide strip stays pure drag surface. */}
             <Sidebar
               open={sidebarOpen}
               onOpen={() => {
